Fix warning.
git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@15248 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ mpi_tests/simple_cm.c
@@ -207,7 +207,7 @@ main(int argc, char* argv[])
  	start = time(NULL);
 	CMCondition_wait(cm, message_wait_condition);
 	end = time(NULL);
-	printf("Elapsed time was %d\n", (end - start));
+	printf("Elapsed time was %d\n", (int)(end - start));
     } else { /* all other processes do this */
 	attr_list contact_list;
 	CMConnection conn = NULL;
